% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{fast.cov}
\alias{fast.cov}
\title{Covariance calculation (speed is a bit improved)}
\usage{
fast.cov(x, y)
}
\arguments{
\item{x}{matrix x}

\item{y}{matrix y}
}
\value{
covariance matrix
}
\description{
Covariance calculation (speed is a bit improved)
}
