% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast.z.cov}
\alias{fast.z.cov}
\title{Covariance z-score calculation (speed is a bit improved)}
\usage{
fast.z.cov(x, y)
}
\arguments{
\item{x}{matrix x}

\item{y}{matrix y}
}
\value{
covariance matrix
}
\description{
Covariance z-score calculation (speed is a bit improved)
}
