% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit.zqtl.adjust}
\alias{fit.zqtl.adjust}
\title{Variational inference of zQTL regression}
\usage{
fit.zqtl.adjust(effect, effect.se, X, multi.C, univar.C)
}
\arguments{
\item{effect}{Marginal effect size matrix (SNP x trait)}

\item{effect.se}{Marginal effect size standard error matrix (SNP x trait)}

\item{X}{Design matrix (reference Ind x SNP)}

\item{n}{sample size of actual data (will ignore if n = 0)}

\item{multi.C}{multivariate SNP confounding factors (SNP x confounder; default: NULL)}

\item{univar.C}{univariate SNP confounding factors (SNP x confounder; default: NULL)}

\item{options}{A list of inference/optimization options.}

\item{do.hyper}{Hyper parameter tuning (default: FALSE)}

\item{do.rescale}{Rescale z-scores by standard deviation (default: FALSE)}

\item{tau}{Fixed value of tau}

\item{pi}{Fixed value of pi}

\item{tau.lb}{Lower-bound of tau (default: -10)}

\item{tau.ub}{Upper-bound of tau (default: -4)}

\item{pi.lb}{Lower-bound of pi (default: -4)}

\item{pi.ub}{Upper-bound of pi (default: -1)}

\item{tol}{Convergence criterion (default: 1e-4)}

\item{gammax}{Maximum precision (default: 1000)}

\item{rate}{Update rate (default: 1e-2)}

\item{decay}{Update rate decay (default: 0)}

\item{jitter}{SD of random jitter for mediation & factorization (default: 0.1)}

\item{nsample}{Number of stochastic samples (default: 10)}

\item{vbiter}{Number of variational Bayes iterations (default: 2000)}

\item{verbose}{Verbosity (default: TRUE)}

\item{print.interv}{Printing interval (default: 10)}

\item{nthread}{Number of threads during calculation (default: 1)}

\item{eigen.tol}{Error tolerance in Eigen decomposition (default: 0.01)}

\item{do.stdize}{Standardize (default: TRUE)}

\item{out.residual}{estimate residual z-scores (default: FALSE)}

\item{nboot}{Number of bootstraps followed by finemapping (default: 100)}

\item{min.se}{Minimum level of SE (default: 1e-4)}

\item{rseed}{Random seed}

\item{scale.var}{Scaled variance calculation (default: FALSE)}
}
\value{
a list of variational inference results.
\itemize{
\item{conf.multi: }{ association with multivariate confounding variables}
\item{conf.uni: }{ association with univariate confounding variables}
\item{resid: }{ residuals}
\item{gwas.clean: }{ cleansed version of univariate GWAS effects}
\item{var: }{ variance decomposition results}
\item{llik: }{ log-likelihood trace over the optimization}
}
}
\description{
Variational inference of zQTL regression
}
\examples{

}
\author{
Yongjin Park, \email{ypp@csail.mit.edu}, \email{yongjin.peter.park@gmail.com}
}
