% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impute.zscore}
\alias{impute.zscore}
\title{Impute z-scores}
\usage{
impute.zscore(Z, V.t, observed)
}
\arguments{
\item{Z}{incomplete z-score matrix}

\item{V.t}{t(V) matrix of the SVD result}

\item{observed}{SNP positions observed in Z}
}
\value{
new z-score matrix
}
\description{
Impute z-scores
}
\details{
Assuming that \eqn{\mathbf{y} = X_{0} \boldsymbol{\theta}_{0} +
X_{1} \boldsymbol{\theta}_{1} \cdots} and \eqn{n^{-1/2}X = U D
V^{\top}}, we have incomplete z-score \eqn{\mathbf{z}_{0} \sim N(V_{0}
D^2 V_{0}^{\top} , V_{0} D^2 V_{0}^{\top})}.  By linear
transformation, we have \deqn{\mathbf{y}_{0} = U D V_{0}^{\top}
\mathbf{\theta}_{0} \sim \mathcal{N}\!\left(U D^{-1}
V_{0}^{\top}\mathbf{z}_{0}, I\right).}  Using this we can derive
imputed z-score: \deqn{\tilde{\mathbf{z}} \approx
VDU^{\top}\mathbf{y}_{0} \sim \mathcal{N}\!\left(V
V_{0}^{\top}\mathbf{z}_{0}, VD^{2}V^{\top}\right).}
}
