% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{log.msg}
\alias{log.msg}
\title{output log message (override `sprintf`)}
\usage{
log.msg(stuff)
}
\arguments{
\item{stuff}{input for sprintf}
}
\description{
output log message (override `sprintf`)
}
\examples{

log.msg("Test Message \%d", 1)

}
