% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{project.zscore}
\alias{project.zscore}
\title{Project z-scores on to different LD block}
\usage{
project.zscore(ld.src.tgt, ld.svd.src, z.src)
}
\arguments{
\item{ld.src.tgt}{SVD result on the target LD block}

\item{ld.svd.src}{SVD result on the source LD block}

\item{z.src}{z-score matrix of the source LD block}

\item{stdize}{standardize if TRUE, otherwise just center the mean}
}
\value{
new z-score matrix
}
\description{
Project z-scores on to different LD block
}
\details{
We assume the estimated individual phenotype:
\deqn{\mathbf{y}_{src} \sim \mathcal{N}\!\left(U_{src}D_{src}^{-1}V_{src}^{\top}\mathbf{z}_{src}, I\right).}
Therefore, a new z-score projected onto the target LD block will be:
\deqn{\mathbf{z}_{tgt} \sim \mathcal{N}\!\left(V_{tgt}D_{tgt}U_{tgt}^{\top}\mathbf{y}_{src}, V_{tgt}D_{tgt}^{2}V_{tgt}^{\top}\right).}
}
