% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{scale.zscore}
\alias{scale.zscore}
\title{Standardize z-scores.}
\usage{
scale.zscore(Z, V.t, D, stdize = TRUE)
}
\arguments{
\item{Z}{z-score matrix}

\item{V.t}{t(V) matrix of the SVD result on genotype matrix}

\item{D}{D diagonal elements of the SVD result on genotype matrix}

\item{stdize}{standardize if TRUE, otherwise just center the mean}
}
\value{
locally standardized z-score matrix
}
\description{
Standardize z-scores.
}
\details{
Estimate \eqn{\mu} and \eqn{\tau} parameters in the model:
\deqn{\mathbf{z} \sim \mathcal{N}\!\left(R (\mu I), \tau^{2}R \right)}
then standardize z-scores by
\deqn{\mathbf{z} \gets (\mathbf{z} - \mu I) / \tau.}
}
