% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{take.ld.pairs}
\alias{take.ld.pairs}
\title{Calculate covariance matrix and transform into triangular form for
visualization}
\usage{
take.ld.pairs(X, cutoff, stdize)
}
\arguments{
\item{X}{genotype matrix}

\item{cutoff}{LD covariance cutoff (default: 0.05)}

\item{stdize}{standardize genotype matrix (default : FALSE)}
}
\value{
\code{data.frame(x, y, x.pos, y.pos, cov)}
}
\description{
Calculate covariance matrix and transform into triangular form for
visualization
}
