% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{take.var.power10}
\alias{take.var.power10}
\title{Estimate standard deviation by the delta method}
\usage{
take.var.power10(var.result)
}
\arguments{
\item{var.result}{which contains two matrix/data.frame:
\itemize{
\item{mean: }{for the mean of log10 values,}
\item{var: }{for the variance of log10 values.}
}}
}
\value{
a list of mean and variance estimation.
\itemize{
\item{mean: }{mean matrix}
\item{sd: }{standard deviation matrix}
}
}
\description{
Estimate standard deviation by the delta method
}
\details{
Say \eqn{x \sim \mathcal{N}\!\left(\mu, \sigma^{2}\right)} in its log10 scale.
Let \eqn{g(x) = 10^x}. Since \eqn{g(x) = \exp(x \log(10))} and
\eqn{g'(x) = \exp(x \log(10)) \log(10)}, by the delta method, we can approximate
the distribution of \eqn{g(x)} around mean value \eqn{\mu}.
\deqn{g(x) \sim \mathcal{N}\!\left(g(\mu), (g'(\mu))^2 \sigma^{2}\right)}.
}
