\name{Collect pairs of LD}
\alias{take.ld.pairs}
\title{Calculate covariance matrix and transform into triangular form}
\description{
}
\usage{
take.ld.pairs(X, cutoff = 0.05, stdize = FALSE)
}
\arguments{
  \item{X}{n x p matrix}
  \item{cutoff}{LD cutoff}
  \item{stdize}{standardize X matrix}
}
\value{
  returns data.frame(x, y, x.pos, y.pos, cov)
}
\details{
}
\references{
}
\examples{
}
\author{Yongjin Park}

