\name{fit.med.zqtl}
\alias{fit.med.zqtl}
\title{Variational inference of zQTL models with two types of summary statistics}
\description{

Measure mediation effects on two types of summary statistics.

}
\usage{
fit.med.zqtl(effect,              # marginal effect : y ~ x
             effect.se,           # marginal se : y ~ x
             effect.m,            # marginal : m ~ x
             effect.m.se,         # marginal se : m ~ x
             X.gwas,              # X matrix for GWAS
             X.med,               # X matrix for mediation
             n = 0,               # sample size of effect
             n.med = 0,           # sample size of effect.m
             C = NULL,            # covariate matrix
             options = list())
}
\arguments{
  \item{effect}{Marginal effect size matrix (SNP x trait)}
  \item{effect.se}{Marginal effect size standard error matrix (SNP x trait)}
  \item{effect.m}{Marignal genetic effects of mediators (SNP x mediator)}
  \item{effect.m.se}{SE of the marginal genetic effects (SNP x mediator)}
  \item{n}{sample size of actual data (will ignore if n = 0)}
  \item{n.med}{sample size of mediation data (will ignore if n = 0)}
  \item{X.gwas}{Design matrix (reference ind.gwas x SNP)}
  \item{X.med}{Design matrix (reference ind.med x SNP)}
  \item{C}{SNP confounding factors (SNP x confounder; default: NULL)}

  \item{options}{A list of inference/optimization options.}
  \item{do.hyper}{Hyper parameter tunign (default: FALSE)}
  \item{do.rescale}{Rescale z-scores by standard deviation (default: TRUE)}
  \item{tau}{Fixed value of tau}
  \item{pi}{Fixed value of pi}

  \item{tau.lb}{Lower-bound of tau (default: -10)}
  \item{tau.ub}{Upper-bound of tau (default: -4)}
  \item{pi.lb}{Lower-bound of pi (default: -4)}
  \item{pi.ub}{Upper-bound of pi (default: -1)}
  \item{tol}{Convergence criterion (default: 1e-4)}
  \item{gammax}{Maximum precision (default: 1000)}
  \item{rate}{Update rate (default: 1e-2)}
  \item{decay}{Update rate decay (default: 0)}
  \item{nsample}{Number of stochastic samples (default: 10)}
  \item{vbiter}{Number of variational Bayes iterations (default: 2000)}
  \item{verbose}{Verbosity (default: TRUE)}
  \item{multivar.mediator}{Multivariate mediator QTL effect (default: FALSE)}

  %% \item{nboot}{Number of bootstrap iterations (default: 100)}
  %% \item{bootstrap.method}{bootstrap with null model, generated by [1]
  %% direct [2] marginal effect (default: 1)}
  \item{direct.model}{1 = average estimated from multiple
    mediators; > 1 = additional effects learn from factorization
    (default: 1)}
  \item{nsingle}{maxumum number of single mediator model
    estimation (default: number of mediators)}
  \item{med.lodds.cutoff}{cutoff for additional direct model factors (default: log-odds > 0)}

  %% \item{med.finemap}{finemapping QTL holding direct effect fixed
  %% (default: FALSE)}

  \item{print.interv}{Printing interval (default: 10)}
  \item{nthread}{Number of threads during calculation (default: 1)}
  \item{eigen.tol}{Error in Eigen/SVD (default: 0.1; fit on 90\% of var)}
  \item{do.stdize}{Standardize (default: TRUE)}
  \item{min.se}{Minimum level of SE}
  \item{rseed}{Random seed}

  \item{weight.m}{Use 1/effect.m.se as weight in the M model (default: TRUE)}
  \item{weight.y}{Use 1/effect.se as weight in the Y model (default: TRUE)}

}
\value{
  \code{fit.zqtl} returns a list of variational parameters
}
\examples{

library(zqtl)

n <- 1000
p <- 2000
n.genes <- 30
h2 <- 0.3
n.causal.direct <- 4
n.causal.qtl <- 3
n.causal.gene <- 2

set.seed(1)

.rnorm <- function(a, b) matrix(rnorm(a * b), a, b)

X.raw <- sapply(1:p, function(j) {
    f <- runif(1, min = 0.1, max = 0.9)
    rbinom(n, 2, f)
})

X <- scale(X.raw)

c.qtls <- matrix(sapply(1:n.genes, function(j) matrix(sample(p, n.causal.qtl))), nrow = n.causal.qtl)

theta.snp.gene <- .rnorm(n.causal.qtl, n.genes) / n.causal.qtl

gene.expr <- lapply(1:dim(c.qtls)[2], function(j) X[, c.qtls[,j], drop = FALSE] \%*\% theta.snp.gene[, j, drop = FALSE] + 0.5 * rnorm(n) )

gene.expr <- do.call(cbind, gene.expr)

theta.med <- sample(c(-1,1), n.causal.gene, TRUE)

causal.direct <- sample(p, n.causal.direct)
theta.dir <- matrix(rnorm(n.causal.direct), n.causal.direct, 1) / n.causal.direct

y <- gene.expr[,1:n.causal.gene,drop = FALSE] \%*\% theta.med # mediation
y <- y + X[, causal.direct, drop = FALSE] \%*\% theta.dir # direct

y <- y + rnorm(n) * c(sqrt(var(y) * (1/h2 - 1)))

################################################################
fast.cov <- function(x, y) {
    n.obs <- crossprod(!is.na(x), !is.na(y))
    ret <- crossprod(replace(x, is.na(x), 0),
                     replace(y, is.na(y), 0)) / n.obs
    return(ret)
}

fast.z.cov <- function(x, y) {
    n.obs <- crossprod(!is.na(x), !is.na(y))
    ret <- crossprod(replace(x, is.na(x), 0),
                     replace(y, is.na(y), 0)) / sqrt(n.obs)
    return(ret)
}

xy.beta <- fast.cov(X, y)
z.xy <- fast.z.cov(X, y)

xy.beta.se <- xy.beta / z.xy

.idx <- sample(n, 200)
X.med <- X[.idx, ] ## smaller sample in mediation QTL

# introduce uncorrected confounders
gene.expr.noisy <- gene.expr + .rnorm(n, 2) \%*\% .rnorm(2, n.genes) / 2

xg.beta <- fast.cov(X.med, gene.expr.noisy[.idx, ])
z.xg <- fast.z.cov(X.med, gene.expr.noisy[.idx, ])
xg.beta.se <- xg.beta / z.xg

################################################################
vb.opt <- list(nsample = 10, vbiter = 3000, rate = 1e-2,
               gammax = 1e4, do.stdize = TRUE, pi = -1, tau = -8,
               do.hyper = FALSE, eigen.tol = 1e-2, tol = 1e-8,
               verbose = TRUE, min.se = 1e-4, re.k = 10,
               print.interv = 200, do.rescale = FALSE)

med.out <- fit.med.zqtl(effect = xy.beta,
                        effect.se = xy.beta.se,
                        effect.m = xg.beta,
                        effect.m.se = xg.beta.se,
                        X.gwas = X,
                        X.med = X.med,
                        options = vb.opt)

plot(med.out$param.mediated$lodds, ylab = 'lodds', main = 'mediation')
plot(med.out$param.unmediated$lodds, ylab = 'lodds', main = 'unmediated')
print(head(med.out$resid.Z))
print(med.out$unmed)
print(med.out$param.intercept)

################################################################
vb.opt <- list(nsample = 10, vbiter = 3000, rate = 1e-2,
               gammax = 1e4, do.stdize = TRUE, pi = -1, tau = -8,
               do.hyper = FALSE, eigen.tol = 1e-2, tol = 1e-8,
               verbose = TRUE, min.se = 1e-4, re.k = 10,
               print.interv = 200, do.rescale = FALSE, direct.model = 10)

med.out <- fit.med.zqtl(effect = xy.beta,
                        effect.se = xy.beta.se,
                        effect.m = xg.beta,
                        effect.m.se = xg.beta.se,
                        X.gwas = X,
                        X.med = X.med,
                        options = vb.opt)

plot(med.out$param.mediated$lodds, ylab = 'lodds', main = 'mediation')
plot(med.out$param.unmediated$lodds, ylab = 'lodds', main = 'unmediated')
print(head(med.out$resid.Z))
print(med.out$unmed)
print(med.out$param.intercept)
}
\details{}
\seealso{}
\keyword{}
