\name{Sparse multivariate QTL modeling on summary statistics}
\alias{zqtl-package}
\alias{zqtl}
\docType{package}
\title{
\packageTitle{zqtl}
}
\description{
\packageDescription{zqtl}
}
\details{}
\author{
\packageAuthor{zqtl}
Maintainer: \packageMaintainer{zqtl}
}
\references{Park, Sarkar, He, Kellis, manuscript in preparation (2017)}
\keyword{package}
\seealso{}
\examples{}
