% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{zscore.pvalue}
\alias{zscore.pvalue}
\title{convert z-score to p-values (two-sided test)}
\usage{
zscore.pvalue(z)
}
\arguments{
\item{z}{z-score}
}
\value{
p-value
}
\description{
convert z-score to p-values (two-sided test)
}
