\name{fit.med.zqtl}
\alias{fit.med.zqtl}
\title{Variational inference of zQTL models with two types of summary statistics}
\description{

Measure mediation effects on two types of summary statistics.

}
\usage{
fit.med.zqtl(effect,              # marginal effect : y ~ x
             effect.se,           # marginal se : y ~ x
             effect.m,            # marginal : m ~ x
             effect.m.se,         # marginal se : m ~ x
             X,                   # X matrix
             n = 0,               # sample size of effect
             n.med = 0,           # sample size of effect.m
             ld = NULL,           # LD matrix
             C = NULL,            # covariate matrix
             options = list())
}
\arguments{
  \item{effect}{Marginal effect size matrix (SNP x trait)}
  \item{effect.se}{Marginal effect size standard error matrix (SNP x trait)}
  \item{effect.m}{Marignal genetic effects of mediators (SNP x mediator)}
  \item{effect.m.se}{SE of the marginal genetic effects (SNP x mediator)}
  \item{n}{sample size of actual data (will ignore if n = 0)}
  \item{n.med}{sample size of mediation data (will ignore if n = 0)}
  \item{X}{Design matrix (reference Ind x SNP)}
  \item{ld}{LD (covariance) matrix (SNP x SNP)}
  \item{C}{SNP confounding factors (SNP x confounder; default: NULL)}

  \item{factored}{Fit factored QTL model (default: FALSE)}
  \item{options}{A list of inference/optimization options.}

  \item{do.hyper}{Hyper parameter tunign (default: FALSE)}
  \item{tau}{Fixed value of tau}
  \item{pi}{Fixed value of pi}

  \item{tau.lb}{Lower-bound of tau (default: -10)}
  \item{tau.ub}{Upper-bound of tau (default: -4)}
  \item{pi.lb}{Lower-bound of pi (default: -4)}
  \item{pi.ub}{Upper-bound of pi (default: -1)}
  \item{tol}{Convergence criterion (default: 1e-4)}
  \item{gammax}{Maximum precision (default: 1000)}
  \item{rate}{Update rate (default: 1e-2)}
  \item{decay}{Update rate decay (default: 0)}
  \item{nsample}{Number of stochastic samples (default: 10)}
  \item{vbiter}{Number of variational Bayes iterations (default: 2000)}
  \item{verbose}{Verbosity (default: TRUE)}
  \item{k}{Rank of the factored model (default: 1)}

  \item{nboot}{Number of bootstrap iterations (default: 100)}

  \item{bootstrap.method}{bootstrap with null model, generated by [1]
  direct [2] marginal effect (default: 1)} 

  \item{med.finemap}{finemapping QTL holding direct effect fixed
  (default: FALSE)}

  \item{med.lodds.cutoff}{mediators included in QTL finemapping
  (default: log-odds > 0)}

  \item{print.interv}{Printing interval (default: 10)}
  \item{nthread}{Number of threads during calculation (default: 1)}
  \item{eigen.tol}{Error in Eigen/SVD (default: 0.1; fit on 90\% of var)}
  \item{do.stdize}{Standardize (default: TRUE)}
  \item{min.se}{Minimum level of SE}
  \item{rseed}{Random seed}

  \item{weight.m}{Use 1/effect.m.se as weight in the M model (default: TRUE)}
  \item{weight.y}{Use 1/effect.se as weight in the Y model (default: TRUE)}
  \item{pretrain}{Pretrain QTL model for fine-tuning (default: FALSE)}
}
\value{
  \code{fit.zqtl} returns a list of variational parameters
}
\examples{

library(zqtl)

n <- 1000
p <- 1000
n.genes <- 20
h2 <- 0.4
n.causal.direct <- 5
n.causal.qtl <- 3
n.causal.gene <- 2

set.seed(1)

X.raw <- sapply(1:p, function(j) {
    f <- runif(1, min = 0.1, max = 0.9)
    rbinom(n, 2, f)
})

X <- as.matrix(apply(X.raw, 2, function(x) x - mean(x, na.rm = TRUE)))

c.qtls <- matrix(sapply(1:n.genes, function(j) matrix(sample(p, n.causal.qtl))), nrow = n.causal.qtl)

theta.snp.gene <- matrix(rnorm(n.causal.qtl * n.genes), n.causal.qtl, n.genes)

gene.expr <- lapply(1:dim(c.qtls)[2], function(j) X[, c.qtls[,j], drop = FALSE] \%*\% theta.snp.gene[, j, drop = FALSE] + 0.5 * rnorm(n) )
gene.expr <- do.call(cbind, gene.expr)

theta.med <- sample(c(-1,1), n.causal.gene, TRUE)

causal.direct <- sample(p, n.causal.direct)
theta.dir <- matrix(rnorm(n.causal.direct), n.causal.direct, 1)

y <- gene.expr[,1:n.causal.gene,drop = FALSE] \%*\% theta.med # mediation
y <- y + X[, causal.direct, drop = FALSE] \%*\% theta.dir # direct

y <- y + rnorm(n) * c(sqrt(var(y) * (1/h2 - 1)))

################################################################
z.xy.tab <- as.matrix(t(apply(X, 2, function(x) coefficients(summary(lm(y ~ x)))[2, 1:2])))
xy.beta <- z.xy.tab[, 1, drop = FALSE]
xy.beta.se <- z.xy.tab[, 2, drop = FALSE]

xg.beta <- matrix(NA, nrow = p, ncol = n.genes)
xg.beta.se <- matrix(NA, nrow = p, ncol = n.genes)

for(g in 1:n.genes) {
    temp <- t(apply(X, 2, function(x) coefficients(summary(lm(gene.expr[, g] ~ x)))[2, 1:2]))
    xg.beta[, g] <- temp[, 1]
    xg.beta.se[, g] <- temp[, 2]
}

vb.opt <- list(nsample = 10, vbiter = 3000, rate = 1e-2, decay = - 0.01, gammax = 1e4,
               ld.matrix = FALSE, do.stdize = FALSE, pi = -0, tau = -4, do.hyper = FALSE,
               eigen.tol = 1e-2, tol = 1e-8, verbose = TRUE, min.se = 1e-4,
               print.interv = 1000, weight.y = FALSE)

med.out <- fit.med.zqtl(effect = xy.beta,
                        effect.se = xy.beta.se,
                        effect.m = xg.beta,
                        effect.m.se = xg.beta.se,
                        X = X, 
                        options = vb.opt)



}
\details{}
\seealso{}
\keyword{}
